+++ src/test/java/com/braintreegateway/integrationtest/TransactionIT.java
@@ -479,6 +479,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         TransactionRequest request = new TransactionRequest().
             amount(TransactionAmount.AUTHORIZE.amount).
             channel("MyShoppingCartProvider").
+            exchangeRateQuoteId("dummyExchangeRateQuoteId-Brainree-Java").
             orderId("123").
             productSku("productsku01").
             creditCard().
@@ -592,6 +593,41 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertEquals("MX", shipping.getCountryCodeAlpha2());
         assertEquals("MEX", shipping.getCountryCodeAlpha3());
         assertEquals("484", shipping.getCountryCodeNumeric());
+    }
+
+    @Test
+    public void saleWithExchangeRateQuoteId() {
+        TransactionRequest request = new TransactionRequest().
+            amount(TransactionAmount.AUTHORIZE.amount).
+            exchangeRateQuoteId("dummyExchangeRateQuoteId-Brainree-Java").
+            creditCard().
+            number(CreditCardNumber.VISA.number).
+            expirationDate("05/2009").
+            done();
+
+        Result<Transaction> result = gateway.transaction().sale(request);
+        assertTrue(result.isSuccess());
+    }
+
+    @Test
+    public void saleErrorWithInvalidExchangeRateQuoteId() {
+        char[] chars = new char[4010];
+        Arrays.fill(chars, 'a');
+        String invalidExchangeRateQuoteId = new String(chars);
+
+        TransactionRequest request = new TransactionRequest().
+            amount(TransactionAmount.AUTHORIZE.amount).
+            exchangeRateQuoteId(invalidExchangeRateQuoteId).
+            creditCard().
+            number(CreditCardNumber.VISA.number).
+            expirationDate("05/2009").
+            done();
+
+        Result<Transaction> result = gateway.transaction().sale(request);
+        assertFalse(result.isSuccess());
+        assertEquals(ValidationErrorCode.EXCHANGE_RATE_QUOTE_ID_TOO_LONG,
+            result.getErrors().forObject("transaction").onField("exchangeRateQuoteId")
+                .get(0).getCode());
     }
 
     @Test